added protobuf and gRPC call

+++ src/Services/Web/Business/Models/ResizeImageRequest.cs
@@ -3,7 +3,7 @@ namespace PixieFit.Web.Business.Models;
 public class ResizeImageRequest
 {
     public byte[] Image { get; set; }
-    public int Width { get; set; }
-    public int Height { get; set; }
-    
+    public decimal Width { get; set; }
+    public decimal Height { get; set; }
+
 }
